+++ app/src/main/java/com/g/vicprojetointegrador/data/model/ReleaseDate.kt
@@ -3,6 +3,5 @@ package com.g.vicprojetointegrador.data.model
 import com.google.gson.annotations.SerializedName
 
 data class ReleaseDate(
-    @SerializedName("certification") val maturityRating: String,
-    @SerializedName("release_date") val releaseDate: String,
+    @SerializedName("certification") val maturityRating: String
 )
