+++ app/src/main/res/layout/activity_generic_error.xml
@@ -14,7 +14,8 @@
         android:backgroundTint="@color/white"
         android:src="@drawable/ic_close"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+        app:layout_constraintTop_toTopOf="parent"
+        android:contentDescription="@string/btnClose_content_desc"/>
 
     <ImageView
         android:id="@+id/ivGenericError"
@@ -24,7 +25,8 @@
         android:background="@drawable/ic_generic_error"
         app:layout_constraintBottom_toTopOf="@+id/tvSystemFailed"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent" />
+        app:layout_constraintStart_toStartOf="parent"
+        android:contentDescription="@string/ivGenericError_content_desc"/>
 
     <TextView
         android:id="@+id/tvSystemFailed"
@@ -40,17 +42,17 @@
 
     <TextView
         android:id="@+id/tvSystemFailedLongText"
+        style="@style/TextAppearance.MovieApp.Body1"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:textAlignment="center"
         android:layout_marginTop="8dp"
         android:lineSpacingExtra="4sp"
         android:text="@string/system_failed_long_desc"
+        android:textAlignment="center"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/tvSystemFailed"
-        style="@style/TextAppearance.MovieApp.Body1">
-    </TextView>
+        tools:ignore="TextContrastCheck"></TextView>
 
     <TextView
         android:id="@+id/tvTryAgain"
@@ -64,7 +66,7 @@
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        tools:ignore="TextContrastCheck" />
+        tools:ignore="TextContrastCheck,TouchTargetSizeCheck" />
 
     <ImageView
         android:id="@+id/ivLine"
