+++ app/src/main/res/values/strings.xml
@@ -24,4 +24,6 @@
     <string name="star_icon_desc">Estrela</string>
     <string name="fab_desc">Retornar para página principal</string>
     <string name="not_found_icon_desc">Lupa de pesquisa não encontrada</string>
+    <string name="btnClose_content_desc">Fechar Tela</string>
+    <string name="ivGenericError_content_desc">Erro na Genérico</string>
 </resources>
