+++ app/src/test/java/com/g/vicprojetointegrador/MovieDetailsUnitTest.kt
@@ -5,11 +5,6 @@ import com.g.vicprojetointegrador.utils.formatPercentage
 import org.junit.Assert.assertEquals
 import org.junit.Test
 
-/**
- * Example local unit test, which will execute on the development machine (host).
- *
- * See [testing documentation](http://d.android.com/tools/testing).
- */
 class MovieDetailsUnitTest {
 
     @Test
